Run jobs in Docker to isolate from host system
From now on, Docker is required on the system
to run FundiesCollab.

Previously, compile and run jobs were run on the
host system meaning it was trivial to remotely
execute anything on the host system.

This removes the remote code execution vulnerability
by running all jobs in an unprivileged container.

Because of this, Docker or some other OCI compatible
container runtime is required to run FundiesCollab.

Signed-off-by: Gary Kim <gary@garykim.dev>